refactor NIST-based decision point into its own namespace & fragment
