+++ data/json/decision_points/nist_800_30/probability_scale_in_5_weighted_levels_ascending_1_0_0.json
@@ -1,0 +1,35 @@
+{
+  "namespace": "nist#800-30",
+  "key": "P_5X",
+  "version": "1.0.0",
+  "name": "Probability Scale in 5 weighted levels, ascending",
+  "definition": "A probability scale with finer resolution at both extremes, based on NIST SP 800-30 Rev. 1 Appendix G",
+  "schemaVersion": "2.0.0",
+  "values": [
+    {
+      "key": "VL",
+      "name": "Very Low",
+      "definition": "0% <= Probability < 5%. Highly unlikely."
+    },
+    {
+      "key": "L",
+      "name": "Low",
+      "definition": "5% <= Probability < 21%. Unlikely."
+    },
+    {
+      "key": "M",
+      "name": "Moderate",
+      "definition": "21% <= Probability < 80%. Somewhat likely."
+    },
+    {
+      "key": "H",
+      "name": "High",
+      "definition": "80% <= Probability < 96%. Highly likely."
+    },
+    {
+      "key": "VH",
+      "name": "Very High",
+      "definition": "96% <= Probability <= 100%. Almost certain."
+    }
+  ]
+}
